#136: BottomSheetInfo 파일 분리
